Changed implementation of service to only use promises.
Data will always be piped in the cache. Cache is valid for a given TTL.
TTL is defined at the top of the service.

If the TTL is valid, the service will not bother checking for new data.
If TTL is not valid, and network is unreachable, data will be served from cache.